requirements.txt
llama-index>=0.12.0
llama-parse>=0.6.0
click>=8.1.0
pydantic>=2.0.0
openai>=1.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
typing-extensions>=4.5.0
aiohttp>=3.8.0
tenacity>=8.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
markdown>=3.4.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0 